Feat/#78 home 지역 수정 api 연동

+++ app/src/main/java/com/paw/key/presentation/ui/home/navigation/HomeLocationSettingNavigation.kt
@@ -20,7 +20,7 @@ fun NavController.navigateHomeLocationSetting(
 fun NavGraphBuilder.homeLocationSettingNavGraph(
     paddingValues: PaddingValues,
     navigateUp: () -> Unit,
-    navigateNext: () -> Unit,
+    navigateNext: (Int) -> Unit,
     navigateHomeLocationSetting: () -> Unit,
     modifier: Modifier = Modifier
 ) {
